Trying to get only useful timeouts with useful error messages
Use promised-timeout in send().

Some calls of send() have to be catched in sds, because the rejections
are not returned and thus lead to uncatched exceptions.

Add optional parameter 'ignoreTimeout' to send().
The server sends no response to the "disconnect" message, so we always
overrun the timeout in this case. To avoid this timeout set parameter
'ignoreTimeout' to true.